MAINT: Structural changes
- Move `relative_l2_error()` from `VerificationUtils` to `VerifcationDataSaving`
- Move `VerifcationDataSaving` from pp.applications.complete_setups.setups_utils.py to pp.viz.data_saving_model_mixin.py
- Update setups and tests accordingly
